% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Provides a descriptive summary of grouped variable.}
\usage{
describe(df, group_var)
}
\arguments{
\item{df}{The data_frame containing the variable to be \code{describe}d.}

\item{group_var}{The variable to be described. Expected to be a bare variable
name.}
}
\value{
A data_frame containing the number of entries in each category, the
    proportion of each category, and the cumulative proportion. The
    data_frame is sorted by descending number of entries.
}
\description{
This function is a convenience function to produce a summary of a grouped
    variable.
}
\examples{
\dontrun{
my_data <- dplyr::data_frame(g1 = c(1, 1, 2, 2, 2))
describe(my_data, g1)
my_data \%>\% describe(g1)
}
}
