% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ess9_labelled}
\alias{ess9_labelled}
\title{Some Labeled Data in the European Social Survey (Round 9)}
\format{
A data frame with 109 observations on the following 4 variables.
\describe{
\item{\code{essround}}{a numeric constant}
\item{\code{edition}}{another numeric constant}
\item{\code{cntry}}{a character vector (with label) for the country in the data}
\item{\code{netusoft}}{a numeric vector (with label) for self-reported internet consumption of a respondent}
}
}
\usage{
ess9_labelled
}
\description{
These are data to illustrate labeled data and how to process them with
\code{get_var_info()} in this package.
}
\details{
Data are condensed summaries from the raw data. They amount to every unique combination of country and
self-reported internet consumption. The data are here to illustrate the \code{get_var_info()} function in this package.
}
\keyword{datasets}
