\name{eustates}
\alias{eustates}
\docType{data}
\title{
EU Member States (Current as of 2019)
}
\description{
European Union membership by accession date
}
\usage{data("eustates")}
\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{\code{date}}{a Date indicating accession}
    \item{\code{country}}{a character vector for country}
    \item{\code{iso2c}}{a character vector for iso2c}
  }
}
\details{

}
\source{
https://europa.eu/european-union/about-eu/countries_en
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(eustates)
## maybe str(eustates) ; plot(eustates) ...
}
\keyword{datasets}
