% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenny.R
\name{jenny}
\alias{jenny}
\title{Set the Only Reproducible Seed That Matters}
\usage{
jenny(x = 8675309)
}
\arguments{
\item{x}{a vector}
}
\value{
When \code{x} is not specified or is 8675309, the function sets a reproducible seed of 8675309 and returns
a nice message congratulating you for it. If \code{x} is not 8675309, the function sets no reproducible seed and
gently admonishes you for wasting its time.
}
\description{
\code{jenny()} sets a reproducible seed of 8675309. It is the only reproducible seed you should use.
}
\details{
\code{jenny()} comes with some additional perks if you have the \pkg{emo} package installed. The package
is optional.
}
\examples{

jenny() # will work and reward you for it
jenny(12345) # will not work and will result in a stern message
}
