% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linloess_plot.R
\name{linloess_plot}
\alias{linloess_plot}
\title{Compare Linear Smoother to LOESS Smoother for Your OLS Model}
\usage{
linloess_plot(mod)
}
\arguments{
\item{mod}{a fitted OLS model}
}
\value{
\code{linloess_plot()} returns a faceted scatterplot as a \pkg{ggplot2} object. You can add cosmetic features
to it after the fact.
}
\description{
\code{linloess_plot()} provides a visual diagnostic of the linearity assumption of the OLS model.
 Provided an OLS model fit by \code{lm()} in base R, the function extracts the model frame and creates a faceted
 scatterplot. For each facet, a linear smoother and LOESS smoother are estimated over the points. Users who run
 this function can assess just how much the linear smoother and LOESS smoother diverge. The more they diverge, the
 more the user can determine how much the OLS model is a good fit as specified. The plot will also point to potential
 outliers that may need further consideration.
}
\details{
This function makes an implicit assumption that there is no variable in the regression
 formula with the name ".y".
}
\examples{

M1 <- lm(mpg ~ ., data=mtcars)

linloess_plot(M1)

}
\author{
Steven V. Miller
}
