% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Multiply a Number by 100 and Round It (By Default: 2)}
\usage{
mround(x, d = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{d}{the number of decimal points to which the user wants to round. If this is not set, it rounds to two decimal points.}
}
\value{
The function takes a numeric vector, multiplies it by 100, rounds it (to two digits by default), and returns it
to the user.
}
\description{
\code{mround()} is a convenience function I wrote for my annotating bar charts that I make.
Assuming a proportion variable, \code{mround()} will multiply each value by 100 and round it for presentation.
By default, it rounds to two. The user can adjust this.
}
\details{
This is a sister function of \code{make_perclab()} in the same package. This, however, won't add a percentage sign.
}
\examples{

x <- runif(100)
mround(x)
mround(x, 2) # same as above
mround(x, 3)
}
