% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2w.R
\name{n2w}
\alias{n2w}
\title{Convert Number (Integer) to Word}
\usage{
n2w(x)
}
\arguments{
\item{x}{a number (one assumes: an integer)}
}
\value{
\code{n2w()} takes a number (one assumes: an integer) and returns that number
as an English word (or set of words).
}
\description{
\code{n2w()} is a convenience function that converts a positive number (and
one assumes an integer) to an English word or set of words.
}
\details{
\code{n2w()} assumes but does not compel the use of integers. Positive reals
are rounded down to the nearest integer.

The function issues a stop in cases where the number provided is negative or
a decimal value between 0 (with precision) and 1. Returns of 0 and infinity
are still valid.

This function has been passed from multiple sources, including John Fox,
Andy Teucher, and an "AJH." I copied it from Jason Miller (no relation) and
made some minor edits to it.
}
\examples{

n2w(3)
n2w(13)
n2w(30)
n2w(8675309)
}
\author{
John Fox, Andy Teucher, "AJH", Jason Miller, Steven V. Miller
}
