% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_refs.R
\name{prepare_refs}
\alias{prepare_refs}
\title{Prepare \pkg{bib2df} Data Frame for Formatting to Various Outputs}
\usage{
prepare_refs(bib2df_refs, toformat = "plain")
}
\arguments{
\item{bib2df_refs}{a data frame created by \pkg{bib2df}}

\item{toformat}{what type of output you are ultimately going to want from \code{print_refs()} . Default is "plain".}
}
\value{
\code{print_refs()}  does some last-minute formatting to a data frame created by
\pkg{bib2df} so that rendering in R Markdown is a little easier and less code-heavy.
}
\description{
\code{prepare_refs} does some last-minute formatting of a data frame created by
\pkg{bib2df} so that it can be formatted nicely to various outputs.
}
\details{
The function is designed to work more generally in the absence of various fields. Assume,
for example, that your data frame has no \code{BOOK} field. The function uses the \code{one_of()} wrapper
to work around this. The "warning" returned by the function is more of a message. This function may be expanded as
I think of more use cases.
}
\examples{

prepare_refs(stevepubs)
}
\seealso{
\code{print_refs()} for formatting a \code{.bib} references to various outputs.
}
