\name{Presidents}
\alias{Presidents}
\docType{data}
\title{
U.S. Presidents and Their Terms in Office
}
\description{
This should be self-evident. Here are all U.S. presidents who have completed their terms in office (i.e. excluding the current one).
}
\usage{data("Presidents")}
\format{
  A data frame with 45 observations on the following 3 variables.
  \describe{
    \item{\code{president}}{a character vector for the name of the president}
    \item{\code{start}}{the start date of the term, as a Date column}
    \item{\code{end}}{the start date of the term, as a Date column}
  }
}
\details{

}
\source{
Scraped, for convenience, from https://www.presidentsusa.net/presvplist.html
}
\references{

}
\examples{
data(Presidents)

}
\keyword{datasets}
