\name{States}
\alias{States}
\docType{data}
\title{
State Abbrevations, State Names, and State Regions/Divisions
}
\description{
A simple data set from \code{state.abb}, \code{state.name}, \code{state.region}, and \code{state.division} (+ District of Columbia). I'd rather just have all these in one place.
}
\usage{data("States")}
\format{
  A data frame with 51 observations on the following 4 variables.
  \describe{
    \item{\code{stateabb}}{a character vector for the state abbreviation}
    \item{\code{statename}}{a character vector for the state name}
    \item{\code{region}}{a character vector for the Census region}
    \item{\code{division}}{a character vector for the Census division}
  }
}
\details{

}
\source{

}
\references{

}
\examples{
data(States)
}
\keyword{datasets}
