% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steveMisc.R
\docType{package}
\name{steveMisc}
\alias{steveMisc}
\alias{steveMisc-package}
\title{\code{steveMisc} package}
\description{
Miscellaneous functions for data science-ing
}
\details{
See the README on
\href{https://github.com/SteveLane/steveMisc#readme}{GitHub}
}
