% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stevepubs}
\alias{stevepubs}
\title{An Incomplete List of My Publications, All of Which You Should Cite}
\format{
A data frame with the following 14 variables.
\describe{
\item{\code{CATEGORY}}{the entry type}
\item{\code{BIBTEXKEY}}{the unique entry key}
\item{\code{AUTHOR}}{a list of authors for this entry}
\item{\code{BOOKTITLE}}{the book title, if appropriate}
\item{\code{JOURNAL}}{the journal title, if appropriate}
\item{\code{NUMBER}}{the journal volume number, if appropriate}
\item{\code{PAGES}}{the range of page numbers, if appropriate}
\item{\code{PUBLISHER}}{the book publisher, if appropriate}
\item{\code{TITLE}}{the title of the publication}
\item{\code{VOLUME}}{the journal volume number, if appropriate}
\item{\code{YEAR}}{the year of publication, as a character. Publications with no year are assumed to be forthcoming}
\item{\code{DOI}}{a DOI, if I entered one}
}
}
\usage{
stevepubs
}
\description{
These are data on my publications, barring a few things like book reviews and some forthcoming pieces.
I use these data to illustrate the \code{print_refs()} function. You should cite my publications more.
}
\details{
Cite my publications more, you goons. *Extremely Smokey Bear voice* Only YOU can jack my h-index to infinity.
}
\keyword{datasets}
