% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tbl_df}
\alias{tbl_df}
\alias{to_tbl}
\title{Convert data frame to an object of class "tibble"}
\usage{
tbl_df(...)

to_tbl(...)
}
\arguments{
\item{...}{optional parameters, but don't put anything here. It's just there to quell CRAN checks.}
}
\value{
This function takes a data frame and turns it into a tibble.
}
\description{
\code{tbl_df()} ensures legacy compatibility with some of my scripts since the function is deprecated in \pkg{dplyr}.
\code{to_tbl()} also added for fun.
}
\examples{

tbl_df(mtcars)
tbl_df(iris)

}
