% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{center_at}
\alias{center_at}
\alias{diff_at}
\alias{r1sd_at}
\alias{r2sd_at}
\alias{within_at}
\title{Scoped Helper Verbs}
\usage{
center_at(data, x, na = TRUE)

diff_at(data, x, l = 1)

r1sd_at(data, x, na = TRUE)

r2sd_at(data, x, na = TRUE)

within_at(data, x, na = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{x}{a vector, likely in your data frame}

\item{na}{what to do with NAs in the vector. Defaults to TRUE
(i.e. passes over the missing observations). Not applicable to \code{diff_at}.}

\item{l}{The order of lags for calculating differences in \code{diff_at}.
Applicable only to \code{diff_at}.}
}
\value{
The function returns a set of new vectors in a data frame after
performing relevant functions. The new vectors have distinct prefixes
corresponding with the action performed on them.
}
\description{
Scoped helper verbs included in this R Documentation file
allow for targeted commands on specified columns. They also rename
the ensuing output to conform to my preferred style. The commands here
are multiple and explained in the details section below.
}
\details{
\code{center_at} is a wrapper for \code{mutate_at} and \code{rename_at} from
\pkg{dplyr}. It takes supplied vectors and effectively centers them from the
mean. It then renames these new variables to have a prefix of \code{c_}. It
is best used \emph{without} a \code{group_by} preceding it. The underlying code is
identical to \code{within_at} in this same package, though it results in new
variables with a different prefix.

\code{r1sd_at} is a wrapper for \code{mutate_at} and \code{rename_at} from
\pkg{dplyr}. It both rescales the supplied vectors to new vectors and renames
the vectors to each have a prefix of \code{s_}. Note the rescaling here is
just by one standard deviation and not two.

\code{r2sd_at} is a wrapper for \code{mutate_at} and \code{rename_at} from
\pkg{dplyr}. It both rescales the supplied vectors to new vectors and renames
the vectors to each have a prefix of \code{z_}. Note the rescaling here is by
two standard deviations and not one.

\code{within_at} is a wrapper for \code{mutate_at} and \code{rename_at} from
\pkg{dplyr}. It takes supplied vectors and effectively centers them from the
mean. It then renames these new variables to have a prefix of \code{w_}. It
is best used with a \code{group_by} preceding it on some kind of "subject",
with an eye toward ultimately including them as covariates in a panel model
to isolate so-called "within" effects.
}
\examples{

center_at(USArrests, c("Murder", "Assault"))

diff_at(USArrests, c("Murder", "Assault"))

r1sd_at(USArrests, c("Murder", "Assault"))

r2sd_at(USArrests, c("Murder", "Assault"))

# Note: best used with a group_by() preceding it.
within_at(USArrests, c("Murder", "Assault"))
}
