% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msl.brms.R
\name{msl.brms}
\alias{msl.brms}
\title{Create Custom \pkg{modelsummary} List from a Model Fitted in \pkg{brms}}
\usage{
msl.brms(mod, robust = FALSE)
}
\arguments{
\item{mod}{a model fitted in \pkg{brms}}

\item{robust}{logical, defaults to FALSE. If FALSE, the "tidy" estimates
are the mean of the simulated coefficients and the standard deviation of those
coefficients. If TRUE, the "tidy" estimates are the median and median absolute
deviation.}
}
\value{
\code{msl.brms()} takes a model fitted by \pkg{brms} and prepares a custom
\pkg{modelsummary} list for formatting with that package's \code{modelsummary()}
function.
}
\description{
\code{msl.brms()} is a convenience function for preparing a model fitted in
\pkg{brms} for presentation in \pkg{modelsummary}  using its custom
"modelsummary_list" feature.
}
\details{
Bayesians typically recoil at the thought of reporting \emph{p}-values for
presentation even as reviewers ask to see these.

Right now, the only easy goodness of fit parameter to return is the number of
observations. The output of this function is easily customized ex post by the
user.
}
\examples{

\dontrun{
# this would take too much time to test...
library(stevedata)
library(brms)

M1 <- brm(gc ~ ool + cvlhs + vl + fl + m + pd, data = Mitchell68,
backend = 'cmdstanr', cores = 4, chains = 4)

msl.brms(M1)

}
}
