\name{States}
\alias{States}
\docType{data}
\title{
State Abbrevations and State Names
}
\description{
A simple data set from \code{state.abb} and \code{state.name} (+ District of Columbia). I'd rather just have all these in one place.
}
\usage{data("States")}
\format{
  A data frame with 51 observations on the following 2 variables.
  \describe{
    \item{\code{stateabb}}{a character vector}
    \item{\code{statename}}{a character vector}
  }
}
\details{

}
\source{

}
\references{

}
\examples{
data(States)
}
\keyword{datasets}
